requirements.txt
ecdsa==0.13.3
pycryptodome==3.9.0
keccak-hash==1.4.3
PyYAML==5.4
eth_account==0.10.0